requirements.txt
Django==3.1.2
django_countries==6.1.3
django_reversion==3.0.8
django_import_export==2.3.0
django_localflavor==3.0.1
factory_boy==3.1.0
coverage==5.3
